Directory refactor (#23)
* Directory structure refactor

* Fix types import

* Change menu keybind to CTRL+K
