+++ src/application_entry.py
@@ -1,0 +1,10 @@
+from text.text_editor import run
+
+
+def main() -> None:
+    """Start the text editor application."""
+    run()
+
+
+if __name__ == "__main__":
+    main()
