+++ src/custom_types/ui_types.py
@@ -1,0 +1,4 @@
+class PopUpDialog:
+    """For type annotation of Dialog box classes"""
+
+    pass
